feat: 계좌 차감 성공 후 월렛 충전 확정 처리

+++ backend/src/main/java/woorifisa/project/backend/domain/wallet/dto/request/ChargeWalletRequest.java
@@ -2,6 +2,6 @@ package woorifisa.project.backend.domain.wallet.dto.request;
 
 public record ChargeWalletRequest(
         Long withdrawAccountId,
-        Long chargeAmount
+        Integer chargeAmount
 ) {
 }
